chore(ui,webui): remove legacy editor code now that PromptFormPage owns the route
Deletes the dead editor surface that the v2 form (#93) supersedes:

- components/ui/src/prompt-editor/{PromptEditorSections,PromptEditorHeader,PromptEditorTabs}.{tsx,stories.tsx} — the cards-stack form (#88/#92), header, and tabs primitive. Type aliases used by the webapp's draft reducer move into prompt-editor/types.ts so import paths through @promptlm/ui keep working.
- components/ui/src/prompts-v2/editor/{cards,cards.stories,PromptEditorTabsLayout}.tsx — the SectionCardVariantProvider chrome wrappers introduced in #88; no consumers after the swap.
- components/ui/src/components/SectionCard/SectionCard.tsx — drops the variant context; SectionCard is back to a single MUI rendering used by non-form surfaces.
- components/promptlm-web-ui/src/features/prompt-editor/PromptEditorPage.tsx + the corresponding promptEditorSections.test.tsx — replaced by PromptFormShell + PromptFormPage in #93.

UI smoke checks adjusted to exercise PromptFormPage instead of the removed PromptEditorHeader/Tabs.

Closes #90 (legacy MUI editor cards removal).
