+++ components/ui/src/prompt-editor/types.ts
@@ -12,6 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+/**
+ * Type aliases used by the legacy editor draft state and the API payload
+ * mapping. The visual editor is now `prompts-v2/form` (`PromptFormPage`),
+ * but the webapp's draft reducer (`features/prompt-editor/draftState.ts`)
+ * and persistence helpers still pivot through these shapes — keeping them
+ * here lets the form shell translate to the new design without forcing a
+ * domain rewrite at the same time.
+ */
+
 import type { ReactNode } from 'react';
 
 export type PromptEditorMode = 'create' | 'edit';
@@ -36,4 +45,89 @@ export type PromptEditorExecutionOption = {
   id: string;
   label: string;
   helperText?: string;
+};
+
+// ── Draft / payload shapes (formerly in PromptEditorSections.tsx) ─────────
+
+export type PromptEditorMessageRole = 'system' | 'user' | 'assistant' | 'tool';
+
+export const DEFAULT_MESSAGE_ROLES: PromptEditorMessageRole[] = [
+  'system',
+  'user',
+  'assistant',
+  'tool',
+];
+
+export type PromptEditorMessage = {
+  id: string;
+  role: PromptEditorMessageRole;
+  content: string;
+  name?: string;
+};
+
+export type PromptEditorPlaceholder = {
+  name: string;
+  value: string;
+};
+
+export type PromptEditorPlaceholderConfig = {
+  startPattern?: string | null;
+  endPattern?: string | null;
+  list: PromptEditorPlaceholder[];
+};
+
+export type PromptEditorToolConfig = {
+  id: string;
+  name: string;
+  scenario: string;
+  mockResponse: string;
+  notes: string;
+};
+
+export type PromptEditorEvaluationPlan = {
+  evaluator: string;
+  type: string;
+  description?: string;
+};
+
+export type PromptEditorEvaluationResult = PromptEditorEvaluationPlan & {
+  success?: boolean;
+  score?: number;
+  reasoning?: string;
+  comments?: string;
+};
+
+export type PromptEditorExecution = {
+  id: string;
+  timestamp: string;
+  response?: {
+    content?: string;
+    usage?: {
+      outputTokens?: number;
+    };
+  };
+  placeholders: { name: string; defaultValue?: string | null }[];
+  evaluations: PromptEditorEvaluationResult[];
+};
+
+export type PromptEditorRequestDraft = {
+  vendor: string;
+  model: string;
+  modelSnapshot?: string | null;
+  url?: string | null;
+  parameters: {
+    temperature?: number;
+    topP?: number;
+    maxTokens?: number;
+    frequencyPenalty?: number;
+    presencePenalty?: number;
+    stream?: boolean;
+  };
+  messages: PromptEditorMessage[];
+};
+
+export type MessageContentSelection = {
+  messageIndex: number;
+  selectionStart: number;
+  selectionEnd: number;
 };
