+++ components/ui/test/ui-package-smoke.mjs
@@ -15,8 +15,7 @@ import {
   Mono,
   PlaceholderToken,
   ProjectSelectionDialog,
-  PromptEditorHeader,
-  PromptEditorTabs,
+  PromptFormPage,
   SectionCard,
   SideNav,
   Sparkline,
@@ -114,46 +113,57 @@ const verifyInfoAndSectionCards = () => {
   assert.match(sectionCardMarkup, /Prompt details/);
 };
 
-const verifyPromptEditorExports = () => {
-  const headerMarkup = renderWithPromptLMTheme(
-    React.createElement(PromptEditorHeader, {
-      mode: 'create',
-      title: 'Create Prompt',
-      description: 'Draft a new prompt definition.',
-      messages: [{ severity: 'info', text: 'Latest draft saved a minute ago.' }],
-      onCreate: () => undefined,
-      onBack: () => undefined,
-    }),
-  );
-  const tabsMarkup = renderWithPromptLMTheme(
-    React.createElement(PromptEditorTabs, {
-      tabs: [
-        { value: 'editor', label: 'Editor' },
-        { value: 'preview', label: 'Preview', badge: 2 },
-        { value: 'history', label: 'History' },
-      ],
-      value: 'preview',
-      onChange: () => undefined,
-      executionOptions: {
-        selectedId: 'execution-2',
-        options: [
-          { id: 'execution-1', label: 'Manual run', helperText: '5 min ago' },
-          { id: 'execution-2', label: 'Canary', helperText: '12 min ago' },
-        ],
-        onSelect: () => undefined,
+const verifyPromptFormPageExport = () => {
+  const draft = {
+    name: 'doc-rag-answer',
+    group: 'rag',
+    description: 'Answer doc questions',
+    request: {
+      type: 'chat',
+      vendor: 'openai',
+      model: 'gpt-4.1',
+      modelSnapshot: '',
+      parameters: {
+        temperature: 0.2,
+        topP: 1,
+        maxTokens: 1024,
+        frequencyPenalty: 0,
+        presencePenalty: 0,
       },
-      tabPanelSlot: React.createElement('div', null, 'Preview content'),
+      messages: [
+        { role: 'system', content: 'You are a helpful assistant.' },
+        { role: 'user', content: 'Hello.' },
+      ],
+    },
+    placeholders: { startPattern: '{{', endPattern: '}}', list: [] },
+    toolConfigs: [],
+    evaluations: [],
+  };
+  const context = {
+    version: '0.1.0',
+    revision: 'r1',
+    repositoryUrl: 'github.com/acme/agents',
+    branch: 'main',
+  };
+  const formMarkup = renderWithPromptLMTheme(
+    React.createElement(PromptFormPage, {
+      mode: 'create',
+      draft,
+      context,
+      evalEnabled: false,
+      onChangeDraft: () => undefined,
+      onToggleEvalEnabled: () => undefined,
+      onCancel: () => undefined,
+      onSaveDraft: () => undefined,
+      onSubmit: () => undefined,
     }),
   );
 
-  assert.match(headerMarkup, /Create Prompt/);
-  assert.match(headerMarkup, /Latest draft saved a minute ago\./);
-  assert.match(headerMarkup, /New prompt/);
-  assert.match(headerMarkup, /Back to prompts/);
-  assert.match(tabsMarkup, /Editor/);
-  assert.match(tabsMarkup, /Preview/);
-  assert.match(tabsMarkup, /Canary/);
-  assert.match(tabsMarkup, /Preview content/);
+  assert.match(formMarkup, /doc-rag-answer/);
+  assert.match(formMarkup, /Save draft/);
+  assert.match(formMarkup, /Identity/);
+  assert.match(formMarkup, /Messages/);
+  assert.match(formMarkup, /github\.com\/acme\/agents/);
 };
 
 const verifyPromptsV2Atoms = () => {
@@ -202,7 +212,7 @@ const verifyPromptsV2Atoms = () => {
 
 verifyLayoutExports();
 verifyInfoAndSectionCards();
-verifyPromptEditorExports();
+verifyPromptFormPageExport();
 verifyPromptsV2Atoms();
 
 console.log('UI package smoke checks passed');
